+++ test/index.js
@@ -21,10 +21,13 @@ import { generateJWK } from '@shgysk8zer0/jwk-utils';
 import { createJWT } from '@shgysk8zer0/jwk-utils/jwt.js';
 import { encrypt, generateSecretKey } from '@shgysk8zer0/aes-gcm';
 import { getSUID } from '@shgysk8zer0/suid';
+import { encodeGeohash } from '@shgysk8zer0/geoutils';
 
 /* global customElements document */
 
 const [jwk, setJWK] = manageState('jwk');
+
+navigator.geolocation.getCurrentPosition(({coords}) => console.log(encodeGeohash(coords)));
 
 generateSecretKey().then(key => encrypt(key, getSUID(), { output: 'base64' })).then(console.log);
 generateJWK().then(({ privateKey }) => createJWT({ iss: 'test' }, privateKey )).then(setJWK).then(console.log(jwk));